add support for stress test
