+++ .github/workflows/main.yml
@@ -125,7 +125,7 @@ jobs:
     - name: Install Go
       uses: actions/setup-go@v2
       with:
-        go-version: 1.19.x
+        go-version: 1.20.x
     - name: Install CGO_ENABLED dependencies
       run: |
         sudo apt-get update && sudo apt-get install -y build-essential
