+++ configs-example/hidra/samples/http_test_google.yml
@@ -4,14 +4,14 @@ tags:
 interval: "1m"
 timeout: "10s"
 steps:
-  - plugin: http
-    action: followRedirects
-    parameters: {}
-  - plugin: http
-    action: request
-    parameters:
-      url: https://google.com/
-  - plugin: http
-    action: statusCodeShouldBe
-    parameters:
-      statusCode: 200
+- plugin: http
+  action: followRedirects
+  parameters: {}
+- plugin: http
+  action: request
+  parameters:
+    url: https://google.com/
+- plugin: http
+  action: statusCodeShouldBe
+  parameters:
+    statusCode: 200
